+++ packages/sit-onyx/src/components/OnyxTable/OnyxTable.vue
@@ -1,10 +1,8 @@
 <script lang="ts" setup>
-import graphSearch from "@sit-onyx/icons/graph-search.svg?raw";
 import { computed, type VNode } from "vue";
 import { useDensity } from "../../composables/density";
 import { injectI18n } from "../../i18n";
 import OnyxEmpty from "../OnyxEmpty/OnyxEmpty.vue";
-import OnyxIcon from "../OnyxIcon/OnyxIcon.vue";
 import type { OnyxTableProps } from "./types";
 
 const props = withDefaults(defineProps<OnyxTableProps>(), {
@@ -29,6 +27,23 @@ const slots = defineSlots<{
    * If unset, a default translated message will be displayed for the current locale.
    */
   empty?(props: { defaultMessage: string }): unknown;
+  /**
+   * Optional slot for showing a headline above the table (top left). See OnyxHeadline component.
+   */
+  headline?(): unknown;
+  /**
+   * Optional slot for showing table actions above the table (top right). See OnyxIconButton and OnyxButton component.
+   */
+  actions?(): unknown;
+  /**
+   * Optional slot for displaying a pagination below the table (bottom right). See OnyxPagination component.
+   */
+  pagination?(): unknown;
+  /**
+   * Optional slot for displaying additional information below the table (bottom left).
+   * Useful for showing a legend, page size selection etc.
+   */
+  bottomLeft?(): unknown;
 }>();
 
 const { t } = injectI18n();
@@ -40,8 +55,21 @@ const isEmptyMessage = computed(() => t.value("table.empty"));
 
 <template>
   <div class="onyx-table-wrapper onyx-component">
+    <div v-if="!!slots.headline || !!slots.actions" class="onyx-table-wrapper__top">
+      <div>
+        <slot name="headline"></slot>
+      </div>
+
+      <div class="onyx-table-wrapper__actions">
+        <slot name="actions"></slot>
+      </div>
+    </div>
+
     <div
-      :class="{ 'onyx-table-wrapper__scroll-container': !props.withPageScrolling }"
+      :class="[
+        'onyx-table-wrapper__container',
+        props.withPageScrolling ? 'onyx-table-wrapper__container--no-scroll' : '',
+      ]"
       :tabindex="props.withPageScrolling ? undefined : 0"
     >
       <table
@@ -82,12 +110,7 @@ const isEmptyMessage = computed(() => t.value("table.empty"));
               <td colspan="100%">
                 <div class="onyx-table__empty-content">
                   <slot name="empty" :default-message="isEmptyMessage">
-                    <OnyxEmpty>
-                      <template #icon>
-                        <OnyxIcon :icon="graphSearch" size="48px" />
-                      </template>
-                      {{ isEmptyMessage }}
-                    </OnyxEmpty>
+                    <OnyxEmpty> {{ isEmptyMessage }} </OnyxEmpty>
                   </slot>
                 </div>
               </td>
@@ -96,13 +119,22 @@ const isEmptyMessage = computed(() => t.value("table.empty"));
         </tbody>
       </table>
     </div>
+
+    <div v-if="!!slots.bottomLeft || !!slots.pagination" class="onyx-table-wrapper__bottom">
+      <div>
+        <slot name="bottomLeft"></slot>
+      </div>
+
+      <div>
+        <slot name="pagination"></slot>
+      </div>
+    </div>
   </div>
 </template>
 
 <style lang="scss">
 @use "../../styles/mixins/layers";
 
-$border-radius: var(--onyx-radius-sm);
 $border: var(--onyx-1px-in-rem) solid var(--onyx-color-component-border-neutral);
 
 /**
@@ -113,55 +145,15 @@ $border: var(--onyx-1px-in-rem) solid var(--onyx-color-component-border-neutral)
 @mixin define-borders() {
   border-spacing: 0;
   border-collapse: separate;
-  border-radius: $border-radius;
 
   // border styles
   th,
   td {
     border-bottom: $border;
-
-    &:first-child {
-      border-left: $border;
-    }
-    &:last-child {
-      border-right: $border;
-    }
   }
 
-  tr:first-of-type th {
-    border-top: $border;
-  }
-
-  // border radius
-  tr:first-child th:first-child {
-    border-top-left-radius: $border-radius;
-  }
-
-  tr:first-child th:last-child {
-    border-top-right-radius: $border-radius;
-  }
-
-  tr:last-child td:first-child {
-    border-bottom-left-radius: $border-radius;
-  }
-
-  tr:last-child td:last-child {
-    border-bottom-right-radius: $border-radius;
-  }
-
-  // special styles if no header exists
-  &:not(:has(thead)) {
-    tr:first-child td {
-      border-top: $border;
-    }
-
-    tr:first-child td:first-child {
-      border-top-left-radius: $border-radius;
-    }
-
-    tr:first-child td:last-child {
-      border-top-right-radius: $border-radius;
-    }
+  tr:last-of-type td {
+    border-bottom: none;
   }
 
   &--vertical-borders {
@@ -176,40 +168,50 @@ $border: var(--onyx-1px-in-rem) solid var(--onyx-color-component-border-neutral)
 
 .onyx-table-wrapper {
   @include layers.component() {
-    position: relative;
+    display: flex;
+    flex-direction: column;
+    gap: var(--onyx-density-xs);
+    font-family: var(--onyx-font-family);
+    color: var(--onyx-color-text-icons-neutral-intense);
 
-    &__scroll-container {
-      border-radius: $border-radius;
+    &__container {
+      border-radius: var(--onyx-radius-sm);
       overflow: auto;
       box-sizing: border-box;
       max-height: inherit;
       max-width: inherit;
       overscroll-behavior: none; // fix bouncy scroll behavior in safari
+      border: $border;
 
       &:focus-visible {
         outline: var(--onyx-outline-width) solid var(--onyx-color-component-focus-primary);
       }
+
+      &--no-scroll {
+        overflow: initial;
+      }
     }
-    // we place a frame on top so the table has visible boundaries
-    // when it is overflowing in the scroll container
-    &::after {
-      content: "";
-      position: absolute;
-      top: 0;
-      left: 0;
-      right: 0;
-      bottom: 0;
-      border: $border;
-      border-radius: $border-radius;
-      pointer-events: none;
+
+    &__top,
+    &__bottom {
+      display: flex;
+      flex-wrap: wrap;
+      align-items: center;
+      justify-content: space-between;
+      gap: var(--onyx-density-xs) var(--onyx-density-xl);
+    }
+
+    &__actions {
+      gap: var(--onyx-density-xs);
+      display: flex;
+      flex-wrap: wrap;
+      align-items: center;
     }
   }
 }
 
 .onyx-table {
   @include layers.component() {
-    font-family: var(--onyx-font-family);
-    color: var(--onyx-color-text-icons-neutral-intense);
     text-align: left;
     contain: paint;
     width: 100%;